+++ demo/Select2Demo.js
@@ -100,9 +100,9 @@ class Demo extends React.Component {
             <div className="demo" >
                 <div id="container">
                 <p>基本使用:</p>
-                <Select defaultValue="lucy" style={{width:200}} onChange={me.handleChange.bind(me)} allowClear={true} getPopupContainer={function() {
+                <Select placeholder="请选择" style={{width:200}} onChange={me.handleChange.bind(me)} allowClear={true} getPopupContainer={function() {
                     return document.getElementById("container")
-                }}>
+                }} dropdownClassName="kuma-select2-selected-has-icon">
                     <Option value="jack">Jack</Option>
                     <Option value="lucy">Lucy</Option>
                     <Option value="disabled" disabled>Disabled</Option>
@@ -116,7 +116,7 @@ class Demo extends React.Component {
                     <Option value="yiminghe">yiminghe</Option>
                 </Select>
                 <p>多选:</p>
-                <Select multiple style={{width:400}} defaultValue={['a10', 'c12']} onChange={me.handleChange.bind(me)}>
+                <Select multiple style={{width:400}} defaultValue={['a10', 'c12']} onChange={me.handleChange.bind(me)} allowClear={true}>
                     {children}
                 </Select>
                 <p>多选提示，提示项根据 ajax 获得</p>
@@ -127,19 +127,6 @@ class Demo extends React.Component {
                 <Select style={{width: '100%'}} searchPlaceholder="标签模式" tags onChange={me.handleChange.bind(me)}>
                     {tagChildren}
                 </Select>
-                {/*<p>分组</p>
-                <Select defaultValue="lucy"
-                    style={{width:200}}
-                    showSearch={false}
-                    onChange={me.handleChange.bind(me)}>
-                    <OptGroup label="Manager">
-                        <Option value="jack">jack</Option>
-                        <Option value="lucy">lucy</Option>
-                    </OptGroup>
-                    <OptGroup label="Engineer">
-                        <Option value="yiminghe">yiminghe</Option>
-                    </OptGroup>
-                </Select>*/}
                 <p>智能提示</p>
                 <Select combobox
                   style={{width:200}}
@@ -155,6 +142,17 @@ class Demo extends React.Component {
                 &nbsp;
                 <Select value={this.state.secondCity} style={{width:150}} onChange={this.onSecondCityChange.bind(this)}>
                     {cityOptions}
+                </Select>
+                <p>Combo 模式</p>
+                <Select combobox={true}>
+                    {cityOptions}
+                </Select>
+                <p>禁用</p>
+                <Select defaultValue="lucy" style={{width:200}} disabled={true}>
+                    <Option value="jack">Jack</Option>
+                    <Option value="lucy">Lucy</Option>
+                    <Option value="disabled" disabled>Disabled</Option>
+                    <Option value="yiminghe">yiminghe</Option>
                 </Select>
                 </div>
             </div>